+++ src/public/types.rs
@@ -149,144 +149,6 @@ pub type MessageSect<'a> =
     Color,   //フォントの色
 );
 
-//マーカーtrait
-pub trait TextUI
-{   fn message( &self ) -> &[ MessageSect ];
-}
-pub trait CountDown
-{   fn initial_count( &self ) -> i32;
-    fn next_state( &self ) -> MyState;
-    fn placeholder( &self ) -> usize;
-    fn to_string( &self, n: i32 ) -> String;
-}
-pub trait HitAnyKey
-{   fn shortcut( &self ) -> MyState;
-}
-
-//TextUIのComponent
-#[derive( Component, Clone, Copy )] pub struct UiStart<'a>
-{   count      : i32,
-    next_state : MyState,
-    message    : &'a [ MessageSect<'a> ],
-    placeholder: usize,
-    string     : fn ( i32 ) -> String,
-}
-impl<'a> TextUI for UiStart<'a>
-{   fn message( &self ) -> &[ MessageSect ] { self.message }
-}
-impl<'a> CountDown for UiStart<'a>
-{   fn initial_count( &self ) -> i32 { self.count + 1 }
-    fn next_state( &self ) -> MyState { self.next_state }
-    fn placeholder( &self ) -> usize { self.placeholder }
-    fn to_string( &self, n: i32 ) -> String { ( self.string )( n ) }
-}
-impl<'a> Default for UiStart<'a>
-{   fn default() -> Self
-    {   Self
-        {   count      : 5,
-            next_state : MyState::MainLoop,
-            message    : UI_START,
-            placeholder: 4,
-            string     : |n| { if n == 0 { "Go!!".to_string() } else { n.to_string() } },
-        }
-    }
-}
-
-#[derive( Component, Clone, Copy )] pub struct UiClear<'a>
-{   count      : i32,
-    next_state : MyState,
-    message    : &'a [ MessageSect<'a> ],
-    placeholder: usize,
-    string     : fn ( i32 ) -> String,
-}
-impl<'a> TextUI for UiClear<'a>
-{   fn message( &self ) -> &[ MessageSect ] { self.message }
-}
-impl<'a> CountDown for UiClear<'a>
-{   fn initial_count( &self ) -> i32 { self.count + 1 }
-    fn next_state( &self ) -> MyState { self.next_state }
-    fn placeholder( &self ) -> usize { self.placeholder }
-    fn to_string( &self, n: i32 ) -> String { ( self.string )( n ) }
-}
-impl<'a> Default for UiClear<'a>
-{   fn default() -> Self
-    {   Self
-        {   count      : 4,
-            next_state : MyState::StageStart,
-            message    : UI_CLEAR,
-            placeholder: 4,
-            string     : |n| { ( n + 6 ).to_string() },
-        }
-    }
-}
-
-#[derive( Component, Clone, Copy )] pub struct UiOver<'a>
-{   count      : i32,
-    next_state : MyState,
-    message    : &'a [ MessageSect<'a> ],
-    placeholder: usize,
-    string     : fn ( i32 ) -> String,
-    shortcut   : MyState,
-}
-impl<'a> TextUI for UiOver<'a>
-{   fn message( &self ) -> &[ MessageSect ] { self.message }
-}
-impl<'a> CountDown for UiOver<'a>
-{   fn initial_count( &self ) -> i32 { self.count + 1 }
-    fn next_state( &self ) -> MyState { self.next_state }
-    fn placeholder( &self ) -> usize { self.placeholder }
-    fn to_string( &self, n: i32 ) -> String { ( self.string )( n ) }
-}
-impl<'a> HitAnyKey for UiOver<'a>
-{   fn shortcut( &self ) -> MyState { self.shortcut }
-}
-impl<'a> Default for UiOver<'a>
-{   fn default() -> Self
-    {   Self
-        {   count      : 10,
-            next_state : MyState::TitleDemo,
-            message    : UI_OVER,
-            placeholder: 6,
-            string     : |n| { n.to_string() },
-            shortcut   : MyState::StageStart,
-        }
-    }
-}
-
-#[derive( Component, Clone, Copy )] pub struct UiTitle<'a>
-{   message    : &'a [ MessageSect<'a> ],
-    shortcut   : MyState,
-}
-impl<'a> TextUI for UiTitle<'a>
-{   fn message( &self ) -> &[ MessageSect ] { self.message }
-}
-impl<'a> HitAnyKey for UiTitle<'a>
-{   fn shortcut( &self ) -> MyState { self.shortcut }
-}
-impl<'a> Default for UiTitle<'a>
-{   fn default() -> Self
-    {   Self
-        {   message    : UI_TITLE,
-            shortcut   : MyState::StageStart,
-        }
-    }
-}
-
-//カウントダウン用のResource
-#[derive( Resource )]
-pub struct CountDownTimer
-{   pub counter: i32,   //カウンター
-    pub timer  : Timer, //１秒タイマー
-}
-impl Default for CountDownTimer
-{   fn default() -> Self
-    {   Self
-        {   counter: 0,
-            timer  : Timer::from_seconds( 1.0, TimerMode::Once ),
-        }
-    }
-}
-
 ////////////////////////////////////////////////////////////////////////////////
 
 //操作を受け付けるgamepadを保存するResource
